@@ -25,12 +25,11 @@
   },
   "dependencies": {
     "is_js": "^0.9.0",
+    "pinia": "2.1.7",
     "swiper": "^5.4.1",
-    "vue": "^2.6.11",
-    "vue-i18n": "^8.18.1",
-    "vue-router": "^3.1.6",
-    "vue-template-compiler": "^2.6.11",
-    "vuex": "^3.4.0"
+    "vue": "3.3.11",
+    "vue-i18n": "9.8.0",
+    "vue-router": "4.2.5"
   },
   "devDependencies": {
     "@babel/core": "^7.9.6",
@@ -48,7 +47,8 @@
     "node-sass": "9.0.0",
     "postcss-loader": "^3.0.0",
     "sass-loader": "13.3.2",
-    "vue-loader": "^15.9.1",
+    "vue-loader": "17.3.1",
+    "vue-style-loader": "^4.1.3",
     "webpack": "5.89.0",
     "webpack-cli": "5.1.4",
     "webpack-dev-server": "4.15.1"
